requirements.txt
streamlit>=1.20
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv
numpy
pymupdf  # optional, remove if not using PDFs
